# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Pharmagarde closer",
        "description": "Get API Pharmagarde Closer",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/gards/position/{lat}/{long}": {
            "get": {
                "tags": [
                    "guardscloser"
                ],
                "summary": "Get json Guards Closer by Lat and Long",
                "description": "Get json Guards Closer by Lat and Long",
                "operationId": "getGuardsCloser",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "path",
                        "description": "Latitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "long",
                        "in": "path",
                        "description": "longitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}